centralized-learning/server/requirements.txt
# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==1.10.13
python-multipart==0.0.6
requests==2.31.0
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ML & Computer Vision
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
facenet-pytorch==2.5.3
numpy==1.26.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Analysis & Visualization
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0